+++ extension/assets/js/content/account.js
@@ -7,17 +7,17 @@ window.Account = (function() {
 	// PRIVATE
 
 	let DEBUG = Debug.ALL.Account,
-		loginPromptsOnThisPage = 0,
-		maxLoginPromptPerPage = 3;
+		dialogueLoginPromptsOnThisPage = 0,
+		maxDialogueLoginPromptsPerPage = 3;
 
 
 
 	/**
 	 *	Return a prompt string to prompt login
 	 */
-	function returnPrompt() {
+	function returnDialogueLoginPrompt() {
 		try {
-			// an array of message prompts
+			// array of potential login prompts
 			let messages = [
 				"Please <a href='" + T.tally_meta.env.website + "/dashboard' target='_blank'>visit your dashboard</a> to connect your account",
 				"You can't stop the trackers unless you <a href='" + T.tally_meta.env.website + "/dashboard' target='_blank'>connect your account</a>",
@@ -30,23 +30,35 @@ window.Account = (function() {
 	}
 
 	/**
-	 *	Tally says a login prompt
+	 *	Check to see if Tally should say a login prompt (and play)
 	 */
-	function checkAndPlayLoginPrompt() {
+	function checkAndPlayDialogueLoginPrompt() {
 		try {
-			if (DEBUG) console.log("🔑 Account.checkAndPlayLoginPrompt() loginPromptsOnThisPage", loginPromptsOnThisPage);
+			const log = "🔑 Account.checkAndPlayDialogueLoginPrompt()";
+			if (DEBUG) console.log(log, "[1.0]",
+				"T.tally_meta.install.loginPrompts.dialogue =", T.tally_meta.install.loginPrompts.dialogue,
+				"dialogueLoginPromptsOnThisPage =", dialogueLoginPromptsOnThisPage
+			);
 
 			let didPrompt = false;
 
 			// if !loggedIn, and we're sure server wasn't just temporarily offline...
 			if (!Page.data.mode.loggedIn &&
-				(T.tally_meta.userLoggedInFailedAttempts >= 5 || T.tally_meta.serverOnlineFailedAttempts >= 5)) {
-				// only play a few times per load
-				if (++loginPromptsOnThisPage > maxLoginPromptPerPage) return;
-				// show dialogue
-				Dialogue.showStr(returnPrompt(), "sad");
+				(T.tally_meta.userLoggedInFailedAttempts >= 5 || T.tally_meta.serverOnlineFailedAttempts >= 5)
+			) {
+				// only play a few times per page
+				if (++dialogueLoginPromptsOnThisPage > maxDialogueLoginPromptsPerPage) return;
+				// don't bother them every time
+				if (T.tally_meta.install.loginPrompts.dialogue % 2 == 0) {
+					setTimeout(function() {
+						Dialogue.showStr(returnDialogueLoginPrompt(), "sad");
+					}, 500);
+				}
 				// return true so calling functions know to stop execution of other game events
 				didPrompt = true;
+				// increment so we know how many times we've attempted
+				T.tally_meta.install.loginPrompts.dialogue++;
+				TallyStorage.saveData("tally_meta", T.tally_meta, log);
 			}
 			return didPrompt;
 
@@ -57,40 +69,9 @@ window.Account = (function() {
 
 
 
-
-
-
-
-
-
-	/**
-	 *	Tally says a login prompt
-	 */
-	function promptHandler() {
-		try {
-
-
-			if (DEBUG) console.log("🔑 Account.promptHandler() ACCOUNT (STILL) BROKEN " +
-				JSON.stringify(T.tally_meta),
-				"T.tally_meta.install.prompts = " + T.tally_meta.install.prompts);
-
-			// don't bother them every time
-			if (T.tally_meta.install.prompts % 2 == 0) {
-				setTimeout(function() {
-					Dialogue.showStr(returnPrompt(), "sad");
-				}, 500);
-			}
-			T.tally_meta.install.prompts++;
-			TallyStorage.saveData("tally_meta", T.tally_meta, "🔑 Account.promptHandler()");
-		} catch (err) {
-			console.error(err);
-		}
-	}
-
-
 	// PUBLIC
 	return {
-		returnPrompt: returnPrompt,
-		checkAndPlayLoginPrompt: checkAndPlayLoginPrompt
+		returnDialogueLoginPrompt: returnDialogueLoginPrompt,
+		checkAndPlayDialogueLoginPrompt: checkAndPlayDialogueLoginPrompt
 	};
 })();